refactor: c++ header file
