+++ include/kernel/board/mmio.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "util.h"
+#include "util.hpp"
 
 #define MMIO_BASE 0x3F000000
