+++ include/kernel/string.hpp
