+++ src/kernel/board/mailbox.cpp
@@ -1,4 +1,4 @@
-#include "board/mailbox.h"
+#include "board/mailbox.hpp"
 
 void mailbox_call(message_t* mailbox) {
   uint32_t data = (((uint32_t)(unsigned long)mailbox) & ~0xf) | 8;
